web: move storage/EncryptionField to its own file
Adapting it to use a core/Field/SettingsField component.

Still pending to extract the storage/EncryptionSettingsDialog and also
to add there a switch to make possible enable/disable the encryption.
