+++ web/src/components/storage/ProposalSettingsSection.test.jsx
@@ -188,150 +188,3 @@ it("requests a volume change when onChange callback is triggered", async () => {
     { volumes: expect.any(Array) }
   );
 });
-
-describe("Encryption field", () => {
-  describe.skip("if encryption password setting is not set yet", () => {
-    beforeEach(() => {
-      // Currently settings cannot be undefined.
-      props.settings = undefined;
-    });
-
-    it("does not render the encryption switch", () => {
-      plainRender(<ProposalSettingsSection {...props} />);
-
-      expect(screen.queryByLabelText("Use encryption")).toBeNull();
-    });
-  });
-
-  describe("if encryption password setting is set", () => {
-    beforeEach(() => {
-      props.settings.encryptionPassword = "";
-    });
-
-    it("renders the encryption switch", () => {
-      plainRender(<ProposalSettingsSection {...props} />);
-
-      screen.getByRole("checkbox", { name: "Use encryption" });
-    });
-  });
-
-  describe("if encryption password is not empty", () => {
-    beforeEach(() => {
-      props.settings.encryptionPassword = "1234";
-    });
-
-    it("renders the encryption switch as selected", () => {
-      plainRender(<ProposalSettingsSection {...props} />);
-
-      const checkbox = screen.getByRole("checkbox", { name: "Use encryption" });
-      expect(checkbox).toBeChecked();
-    });
-
-    it("renders a button for changing the encryption settings", () => {
-      plainRender(<ProposalSettingsSection {...props} />);
-
-      screen.getByRole("button", { name: /Encryption settings/ });
-    });
-
-    it("changes the selection on click", async () => {
-      const { user } = plainRender(<ProposalSettingsSection {...props} />);
-
-      const checkbox = screen.getByRole("checkbox", { name: "Use encryption" });
-      await user.click(checkbox);
-
-      expect(checkbox).not.toBeChecked();
-      expect(props.onChange).toHaveBeenCalled();
-    });
-
-    it("allows changing the encryption settings when clicking on the settings button", async () => {
-      const { user } = plainRender(<ProposalSettingsSection {...props} />);
-
-      const button = screen.getByRole("button", { name: /Encryption settings/ });
-      await user.click(button);
-
-      const popup = await screen.findByRole("dialog");
-      screen.getByText("Encryption settings");
-
-      const accept = within(popup).getByRole("button", { name: "Accept" });
-      await user.click(accept);
-
-      expect(screen.queryByRole("dialog")).not.toBeInTheDocument();
-      expect(props.onChange).toHaveBeenCalled();
-    });
-
-    it("allows canceling the changes of the encryption settings", async () => {
-      const { user } = plainRender(<ProposalSettingsSection {...props} />);
-
-      const button = screen.getByRole("button", { name: /Encryption settings/ });
-      await user.click(button);
-
-      const popup = await screen.findByRole("dialog");
-      screen.getByText("Encryption settings");
-
-      const cancel = within(popup).getByRole("button", { name: "Cancel" });
-      await user.click(cancel);
-
-      expect(screen.queryByRole("dialog")).not.toBeInTheDocument();
-      expect(props.onChange).not.toHaveBeenCalled();
-    });
-  });
-
-  describe("if encryption password is empty", () => {
-    beforeEach(() => {
-      props.settings.encryptionPassword = "";
-    });
-
-    it("renders the encryption switch as not selected", () => {
-      plainRender(<ProposalSettingsSection {...props} />);
-
-      const checkbox = screen.getByRole("checkbox", { name: "Use encryption" });
-      expect(checkbox).not.toBeChecked();
-    });
-
-    it("does not render a button for changing the encryption settings", () => {
-      plainRender(<ProposalSettingsSection {...props} />);
-
-      const button = screen.queryByRole("button", { name: /Encryption settings/ });
-      expect(button).toBeNull();
-    });
-
-    it("changes the selection and allows changing the settings on click", async () => {
-      const { user } = plainRender(<ProposalSettingsSection {...props} />);
-
-      const checkbox = screen.getByRole("checkbox", { name: "Use encryption" });
-      await user.click(checkbox);
-
-      const popup = await screen.findByRole("dialog");
-      screen.getByText("Encryption settings");
-
-      const passwordInput = screen.getByLabelText("Password");
-      const passwordConfirmInput = screen.getByLabelText("Password confirmation");
-      await user.type(passwordInput, "1234");
-      await user.type(passwordConfirmInput, "1234");
-      const accept = within(popup).getByRole("button", { name: "Accept" });
-      await user.click(accept);
-
-      expect(screen.queryByRole("dialog")).not.toBeInTheDocument();
-
-      expect(props.onChange).toHaveBeenCalled();
-      expect(checkbox).toBeChecked();
-    });
-
-    it("does not select encryption if the settings are canceled", async () => {
-      const { user } = plainRender(<ProposalSettingsSection {...props} />);
-
-      const checkbox = screen.getByRole("checkbox", { name: "Use encryption" });
-      await user.click(checkbox);
-
-      const popup = await screen.findByRole("dialog");
-      screen.getByText("Encryption settings");
-
-      const cancel = within(popup).getByRole("button", { name: "Cancel" });
-      await user.click(cancel);
-
-      expect(screen.queryByRole("dialog")).not.toBeInTheDocument();
-      expect(props.onChange).not.toHaveBeenCalled();
-      expect(checkbox).not.toBeChecked();
-    });
-  });
-});
